+++ cli/main.go
@@ -1,7 +1,16 @@
 package main
 
-import "github.com/preview-manager/cli/cmd"
+import (
+	_ "embed"
+	"strings"
+
+	"github.com/preview-manager/cli/cmd"
+)
+
+//go:embed VERSION
+var version string
 
 func main() {
+	cmd.SetVersion(strings.TrimSpace(version))
 	cmd.Execute()
 }